declares announcements in path
