+++ plugin/announcement/assets.json
@@ -1,7 +1,7 @@
 {
   "webpack": {
     "entry": {
-      "announcement-resource": "resources/announcement"
+      "announcement-resource": "resources/announcement/bootstrap"
     }
   },
